foodgram/requirements.txt
django
djangorestframework
pillow
django-debug-toolbar
gunicorn
psycopg2-binary